Moved OperationResult class to coreAPI directory.
