+++ src/core/API/OperationResult.cpp
